feat(wdk rules) support clang and llvm for wdk rules

+++ tests/projects/windows/driver/kmdf/ioctl/exe/testapp.c
@@ -213,7 +213,7 @@ GetCoinstallerVersion(
 
 VOID __cdecl
 main(
-    _In_ ULONG argc,
+    _In_ INT argc,
     _In_reads_(argc) PCHAR argv[]
     )
 {
